@@ -1,8 +1,9 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {MatSidenav} from '@angular/material/sidenav';
 import {Router} from '@angular/router';
 import {TreeviewService} from '../services/treeview.service';
 import {Treeview} from '../models/treeview';
-import * as _ from 'underscore';
+import * as _ from 'lodash';
 
 @Component({
   selector: 'app-content',
@@ -10,7 +11,8 @@ import * as _ from 'underscore';
   styleUrls: ['./content.component.css']
 })
 export class ContentComponent implements OnInit {
-  treeviewPanel: Treeview = new Treeview;
+  @ViewChild('sidenav') sidenav: MatSidenav;
+  treeviewPanel: Treeview = new Treeview();
 
   constructor(private route: Router, private treeviewService: TreeviewService) {
     this.treeviewService.getTreeview().then(
@@ -19,6 +21,7 @@ export class ContentComponent implements OnInit {
           treeviews,
           treeview => treeview.url == this.route.url
         );
+        this.treeviewPanel = this.treeviewPanel || new Treeview();
       }
     );
   }
