@@ -1,12 +1,15 @@
-import { NgModule } from '@angular/core';
+import {NgModule} from '@angular/core';
 
-import { CommonModule } from '@angular/common';
-import { ContentRoutingModule} from './content-routing.module';
-import { MainHeaderModule } from '../main-header/main-header.module';
-import { MainSidebarModule } from '../main-sidebar/main-sidebar.module';
-import { DashboardModule } from '../dashboard/dashboard.module';
+import {CommonModule} from '@angular/common';
+import {RouterModule} from '@angular/router';
+import {MainHeaderModule} from '../main-header/main-header.module';
+import {MainSidebarModule} from '../main-sidebar/main-sidebar.module';
+import {DashboardModule} from '../dashboard/dashboard.module';
+import {MaterialModule} from '../material/material.module';
 
-import { ContentComponent } from './content.component';
+import {ContentComponent} from './content.component';
+import {ErrorPageComponent} from "./error-page/error-page.component";
+import {TitlePanelComponent} from './title-panel/title-panel.component';
 
 @NgModule({
   imports: [
@@ -14,9 +17,11 @@ import { ContentComponent } from './content.component';
     MainHeaderModule,
     MainSidebarModule,
     DashboardModule,
-    ContentRoutingModule
+    RouterModule,
+    MaterialModule
   ],
-  declarations: [ContentComponent],
+  declarations: [ContentComponent, ErrorPageComponent, TitlePanelComponent],
   exports: [ContentComponent]
 })
-export class ContentModule { }
+export class ContentModule {
+}
