@@ -1,4 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {AuthResponse, FacebookService, LoginStatus} from "ngx-facebook";
+import {TransverseData} from "../../models/constants/transverse-data";
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-user-menu',
@@ -7,9 +10,26 @@ import { Component, OnInit } from '@angular/core';
 })
 export class UserMenuComponent implements OnInit {
 
-  constructor() { }
+  constructor(private fb: FacebookService, private router: Router) {
+  }
 
   ngOnInit() {
   }
 
+  logout(): void {
+    let accessToken = localStorage.getItem(TransverseData.accessToken);
+    this.fb.getLoginStatus().then(
+      (loginStatus: LoginStatus) => {
+        if (loginStatus.status === 'connected') {
+          this.fb.logout().then(
+            (authResponse: AuthResponse) => {
+              localStorage.removeItem(TransverseData.accessToken);
+              localStorage.removeItem(TransverseData.currentUser);
+              this.router.navigate(['login']);
+            }
+          );
+        }
+      }
+    );
+  }
 }
