@@ -1,8 +1,10 @@
 export class User {
   _id: string;
+  first_name: string;
+  last_name: string;
+  email: string;
+  picture: string;
   username: string;
-  firstname: string;
-  lastname: string;
   job: string;
   password: string;
   role: string;
