@@ -1,18 +1,19 @@
 import {Injectable} from '@angular/core';
-import {Headers, Http, Response} from '@angular/http';
+import {HttpClient} from '@angular/common/http';
+import {Treeview} from '../models/treeview';
 
 @Injectable()
 export class TreeviewService {
   private getUserUrl = 'http://localhost:3061/api/treeview';
 
-  constructor(private http: Http) {
+  constructor(private http: HttpClient) {
   }
 
   getTreeview() {
     return this.http
       .get(this.getUserUrl)
       .toPromise()
-      .then(response => response.json())
+      .then(response => response as Treeview)
       .catch(this.handleError);
   }
 
@@ -20,7 +21,24 @@ export class TreeviewService {
     return this.http
       .post(this.getUserUrl, treeview)
       .toPromise()
-      .then(response => response.json())
+      .then(response => response as Treeview)
+      .catch(this.handleError);
+  }
+
+  updateTreeview(treeview: Treeview) {
+    return this.http
+      .put(`${this.getUserUrl}/${treeview._id}`, treeview)
+      .toPromise()
+      .then(response => response as Treeview)
+      .catch(this.handleError);
+  }
+
+  // Implement a method to get the public deals
+  deleteTreeview(id: string) {
+    return this.http
+      .delete(`${this.getUserUrl}/${id}`)
+      .toPromise()
+      .then(response => response as Treeview)
       .catch(this.handleError);
   }
 
