@@ -1,8 +1,8 @@
-import {Component, Input, OnInit} from '@angular/core';
-import {MatDialog} from '@angular/material';
+import {Component, Input, Output, OnInit, EventEmitter} from '@angular/core';
+import {MatDialog, MatDialogConfig} from '@angular/material';
 import {UserService} from '../../services/user.service';
 import {UserModalComponent} from '../user-modal/user-modal.component';
-import {User} from "../../models/user";
+import {User} from '../../models/user';
 
 @Component({
   selector: 'app-user-item',
@@ -10,7 +10,8 @@ import {User} from "../../models/user";
   styleUrls: ['./user-item.component.css']
 })
 export class UserItemComponent implements OnInit {
-  @Input() user: any;
+  @Input() user: User;
+  @Output() onDelete: EventEmitter<any> = new EventEmitter();
 
   constructor(private userService: UserService,
               public dialog: MatDialog) {
@@ -20,10 +21,12 @@ export class UserItemComponent implements OnInit {
   }
 
   openDialog(): void {
-    const dialogRef = this.dialog.open(UserModalComponent, {
+    const dialogConfig = {
       width: '500px',
       data: this.user
-    });
+    };
+
+    const dialogRef = this.dialog.open(UserModalComponent, dialogConfig as MatDialogConfig);
 
     dialogRef.afterClosed().subscribe(result => {
       if (!!result) {
@@ -32,7 +35,7 @@ export class UserItemComponent implements OnInit {
     });
   }
 
-  delete(){
-    this.userService.deleteUser(this.user._id).then(() => this.user = new User());
+  deleteUser() {
+    this.userService.deleteUser(this.user._id).subscribe(() => this.onDelete.emit());
   }
 }
