@@ -1,17 +1,13 @@
-import {Component, OnInit, Input} from '@angular/core';
+import {Component, Input} from '@angular/core';
 
 @Component({
   selector: 'app-information-badge',
   templateUrl: './information-badge.component.html',
   styleUrls: ['./information-badge.component.css']
 })
-export class InformationBadgeComponent implements OnInit {
+export class InformationBadgeComponent {
   @Input() data: any;
 
   constructor() {
   }
-
-  ngOnInit() {
-  }
-
 }
