@@ -1,12 +1,15 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { InformationBadgeComponent } from './information-badge/information-badge.component';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {MaterialModule} from '../material/material.module';
+import {InformationBadgeComponent} from './information-badge/information-badge.component';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    MaterialModule
   ],
   declarations: [InformationBadgeComponent],
   exports: [InformationBadgeComponent]
 })
-export class WidgetModule { }
+export class WidgetModule {
+}
